requirements.txt
langgraph
langchain-openai
langchain-core
# transformers==4.33.0
func_timeout
sqlparse
requests
pandas
tiktoken
tqdm
nltk
flask
# EvoSQL dependencies
pyyaml
openai
